LayoutTests/svg/dynamic-updates/SVGFEFloodElement-dom-flood-opacity-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'flood-opacity' attribute of the SVGFEFloodElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS floodElement.getAttribute('flood-opacity') is ".1"
PASS floodElement.getAttribute('flood-opacity') is ".8"
PASS successfullyParsed is true

TEST COMPLETE
